contact _us changes

+++ src/main/java/com/londonzone/repositories/ContactUsRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.londonzone.domain.ContactUs;;
 
 @Repository
-public interface ContactUsRepository extends CrudRepository<ContactUs, Integer> {
+public interface ContactUsRepository extends CrudRepository<ContactUs, Long> {
 
 }
